In 3.0.0, we provide one MongoDB module: `log4j-mongodb`, based on the (#2493)
current MongoDB Client Driver.

We renamed the `log4j-mongodb4`, module, package, and classes in favor
of `log4j-mongodb`, in other words, drop the `4`.

We no longer provide the modules `log4j-mongodb2`, `log4j-mongodb3`,
`log4j-mongodb4`.
