--- log4j-mongodb4/src/test/java/org/apache/logging/log4j/mongodb4/MongoDb4ResolverTest.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to you under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.logging.log4j.mongodb4;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoIterable;
-import org.junit.jupiter.api.Test;
-
-/**
- * Tests MongoDbRule.
- * <p>
- * The test framework {@code de.flapdoodle.embed.mongo} requires Java 8.
- * </p>
- */
-@UsingMongoDb4
-public class MongoDb4ResolverTest {
-
-    @Test
-    public void testAccess(final MongoClient mongoClient) {
-        final MongoIterable<String> databaseNames = mongoClient.listDatabaseNames();
-        assertNotNull(databaseNames);
-        assertNotNull(databaseNames.first());
-    }
-}
