+++ pom.xml
@@ -254,7 +254,7 @@
     <module>log4j-kit</module>
     <module>log4j-layout-template-json</module>
     <module>log4j-layout-template-json-test</module>
-    <module>log4j-mongodb4</module>
+    <module>log4j-mongodb</module>
     <module>log4j-osgi-test</module>
     <module>log4j-perf-test</module>
     <module>log4j-plugin-processor</module>
@@ -494,7 +494,7 @@
 
       <dependency>
         <groupId>org.apache.logging.log4j</groupId>
-        <artifactId>log4j-mongodb4</artifactId>
+        <artifactId>log4j-mongodb</artifactId>
         <version>${project.version}</version>
       </dependency>
