+++ src/site/antora/modules/ROOT/pages/manual/messages.adoc
@@ -225,7 +225,7 @@ with a no layout, it converts a Log4j `MapMessage` to a JMS
 * When a xref:manual/appenders.adoc#JDBCAppender[JDBC Appender] is configured
 with no layout, it converts a Log4j `MapMessage` to values in a
 SQL INSERT statement.
-* When a xref:manual/appenders.adoc#NoSQLAppenderMongoDB4[MongoDB4 Appender] is
+* When a xref:manual/appenders.adoc#NoSQLAppenderMongoDB[MongoDB Appender] is
 configured with no layout, it converts a Log4j `MapMessage` to
 fields in a MongoDB object.
